// src/api/types.ts
export interface StatisticsDataItem {
  ORG_ID: string; // 기관코드
  TBL_ID: string; // 통계표ID
  TBL_NM: string; // 통계표명
  C1?: string; // 분류값 ID1
  C1_OBJ_NM?: string; // 분류명1
  C1_NM?: string; // 분류값명1
  C2?: string; // 분류값 ID2
  C2_OBJ_NM?: string;
  C2_NM?: string;
  C3?: string;
  C3_OBJ_NM?: string;
  C3_NM?: string;
  C4?: string;
  C4_OBJ_NM?: string;
  C4_NM?: string;
  C5?: string;
  C5_OBJ_NM?: string;
  C5_NM?: string;
  C6?: string;
  C6_OBJ_NM?: string;
  C6_NM?: string;
  C7?: string;
  C7_OBJ_NM?: string;
  C7_NM?: string;
  C8?: string;
  C8_OBJ_NM?: string;
  C8_NM?: string;
  ITM_ID: string; // 항목 ID
  ITM_NM: string; // 항목명
  ITM_NM_ENG?: string; // 항목영문명
  UNIT_ID?: string; // 단위ID
  UNIT_NM: string; // 단위명
  UNIT_NM_ENG?: string; // 단위영문명
  PRD_SE: string; // 수록주기
  PRD_DE: string; // 수록시점
  DT: string; // 수치값
  LST_CHN_DE?: string; // 최종수정일
}
// ...
export interface SimplifiedDataItem {
  period: string;
  rawPeriod?: string;
  periodType?: string;
  value: string;
  rawValue?: string;
  classification: string;
  classificationCodes: Record<string, string | undefined>;
  item: string;
  itemId?: string;
  unit: string;
  unitId?: string;
  raw: StatisticsDataItem;
}